changed smith to smelt and smithed to smith
