+++ src/commands/Minion/minion.ts
@@ -57,7 +57,7 @@ export default class MinionCommand extends BotCommand {
 			cooldown: 1,
 			aliases: ['m'],
 			usage:
-				'[clues|k|kill|setname|buy|clue|kc|pat|stats|mine|smith|quest|qp|chop|ironman|light|fish|laps|cook|smithed] [quantity:int{1}|name:...string] [name:...string]',
+				'[clues|k|kill|setname|buy|clue|kc|pat|stats|mine|smith|quest|qp|chop|ironman|light|fish|laps|cook|smelt] [quantity:int{1}|name:...string] [name:...string]',
 
 			usageDelim: ' ',
 			subcommands: true
@@ -381,9 +381,9 @@ ${Emoji.QuestIcon} QP: ${msg.author.settings.get(UserSettings.QP)}
 			});
 	}
 
-	async smith(msg: KlasaMessage, [quantity, barName]: [number, string]) {
+	async smelt(msg: KlasaMessage, [quantity, barName]: [number, string]) {
 		await this.client.commands
-			.get('smith')!
+			.get('smelt')!
 			.run(msg, [quantity, barName])
 			.catch(err => {
 				throw err;
@@ -400,6 +400,9 @@ ${Emoji.QuestIcon} QP: ${msg.author.settings.get(UserSettings.QP)}
 		}
 	async smithed(msg: KlasaMessage, [quantity, smithedBarName]: [number, string]) {
 		this.client.commands.get('smithed')!.run(msg, [quantity, smithedBarName]);
+	}
+	async smith(msg: KlasaMessage, [quantity, smithedBarName]: [number, string]) {
+		this.client.commands.get('smith')!.run(msg, [quantity, smithedBarName]);
 	}
 
 	async chop(msg: KlasaMessage, [quantity, logName]: [number, string]) {
